after compiling with Wall

+++ c++/003.cpp
@@ -9,26 +9,26 @@
 int main() {
   const uint64_t number = 600851475143;
 
-  int max_prime = 0;
-  int i;
+  uint64_t max_prime = 0;
+  uint64_t i;
   for (i = 2; i < sqrt(number); ++i) {
     int matched_2 = 0;
     int matched_3 = 0;
     uint64_t divi = 0;
     if (!(number % i)) {
       divi = number/i;
-      printf("FACTOR of %lu is (%d * %lu)\n", number, i, divi);
-      int j;
+      printf("FACTOR of %lu is (%lu * %lu)\n", number, i, divi);
+      uint64_t j;
       for (j = 2; j < sqrt(i); j++) {
         if (!(i % j)) {
-          printf("2 FACTOR of %d is (%d * %d)\n", i, j, i/j);
+          printf("2 FACTOR of %lu is (%lu * %lu)\n", i, j, i/j);
           matched_2 = 1;
         }
       }
-      int k;
+      uint64_t k;
       for (k = 2; k < sqrt(divi); k++) {
         if (!(divi % k)) {
-          printf("3 FACTOR of %lu is (%d * %lu)\n", divi, k, divi/k);
+          printf("3 FACTOR of %lu is (%lu * %lu)\n", divi, k, divi/k);
           matched_3 = 1;
         }
       }
@@ -40,6 +40,6 @@ int main() {
         max_prime = divi;
     }
   }
-  printf("MAX PRIME DIVISOR of %lu is %d\n", number, max_prime);
+  printf("MAX PRIME DIVISOR of %lu is %lu\n", number, max_prime);
   return(0);
 }
